Extract route handlers
